move ValueType to enum class and fix tests from OptionSkip

+++ db_stress_tool/cf_consistency_stress.cc
@@ -757,7 +757,7 @@ class CfConsistencyStressTest : public StressTest {
             for (const KeyVersion& kv : versions) {
               fprintf(stderr, "  key %s seq %" PRIu64 " type %d\n",
                       Slice(kv.user_key).ToString(true).c_str(), kv.sequence,
-                      kv.type);
+                      (int)kv.type);
             }
           };
